+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## Improvements
 
 - PR #214 Add codeowners
+- PR #226 Add some tests of the Python `DeviceBuffer`
 
 ## Bug Fixes